+++ apps/stripe/src/app/api/saleor/payment-gateway-initialize-session/response-shape.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { PaymentGatewayInitializeResponseShape } from "@/app/api/saleor/gateway-initialize/response-shape";
+import { PaymentGatewayInitializeResponseShape } from "@/app/api/saleor/payment-gateway-initialize-session/response-shape";
 
 describe("PaymentGatewayInitializeResponseShape", () => {
   it("Constructs with valid pk value", () => {
