+++ apps/stripe/src/app/api/saleor/payment-gateway-initialize-session/response-shape.ts
